refactor: EpicEditModal 컴포넌트 수정 [#186]
에픽을 수정하기 위해 사용된 모달을 재사용하기 위해 시그니처 변경
