+++ client/src/components/EpicEditModal/index.tsx
@@ -1,21 +1,17 @@
 import React, { useEffect, useState } from 'react';
 import S from './style';
 import { EpicType } from '@/types/epic';
-import { getYMD, isLatter, isSameDay } from '@/lib/utils/date';
+import { getYMD } from '@/lib/utils/date';
 import { Modal } from '@/lib/design';
-import { useEpicDispatch } from '@/lib/hooks/useContextHooks';
-import { updateEpicById } from '@/lib/api/epic';
-import { useSocketSend } from '@/lib/hooks';
-import { toast } from 'react-toastify';
-import { errorMessage } from '@/lib/common/message';
-import { checkStringInput } from '@/lib/utils/bytes';
 
 interface EpicEditModalProps {
+  title?: string;
   showEditModal: boolean;
   setShowModal: React.Dispatch<React.SetStateAction<boolean>>;
   epicData: EpicType;
   value: string;
   handleChange: (e: React.ChangeEvent) => void;
+  handleFormSubmit: (e: React.FormEvent, datePair: { startDate: Date; endDate: Date }) => void;
 }
 
 const formatDate = (date: Date) => {
@@ -24,53 +20,31 @@ const formatDate = (date: Date) => {
 };
 
 const EpicEditModal = ({
+  title,
   showEditModal,
   setShowModal,
   epicData,
   value,
   handleChange,
+  handleFormSubmit,
 }: EpicEditModalProps) => {
   const [startDate, setStartDate] = useState(epicData.startAt);
   const [endDate, setEndDate] = useState(epicData.endAt);
-  const dispatchEpic = useEpicDispatch();
-  const emitUpdateEpic = useSocketSend('UPDATE_EPIC_BAR');
 
   useEffect(() => {
     setStartDate(epicData.startAt);
     setEndDate(epicData.endAt);
   }, [epicData]);
 
-  const handleFormSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!value.length || !checkStringInput(value)) {
-      toast.error(errorMessage.EPIC_TITLE_LENGTH_LIMIT);
-      return;
-    }
-
-    if (isLatter(startDate, endDate) && !isSameDay(startDate, endDate)) {
-      toast.error(errorMessage.START_DATE_IS_LATTER);
-      return;
-    }
-    const updatedEpic = { ...epicData, name: value, startAt: startDate, endAt: endDate };
-    dispatchEpic({
-      type: 'UPDATE_EPIC',
-      epic: updatedEpic,
-    });
-    setShowModal(false);
-
-    await updateEpicById(epicData.id, updatedEpic);
-    emitUpdateEpic(epicData.id);
-  };
-
   return (
     <Modal
       shouldConfirm
-      title="에픽 정보 수정"
+      title={title ?? '에픽 정보 수정'}
       visible={showEditModal}
       onClose={() => setShowModal(false)}
-      onClickOk={handleFormSubmit}
+      onClickOk={(e) => handleFormSubmit(e, { startDate, endDate })}
     >
-      <S.Form onSubmit={handleFormSubmit}>
+      <S.Form onSubmit={(e) => handleFormSubmit(e, { startDate, endDate })}>
         <S.Input
           type="text"
           placeholder={'에픽 제목을 입력하세요 (255자 이내)'}
